Removes specific bytes size from test.

+++ test/test_util.py
@@ -373,13 +373,13 @@ class TestUnit(unittest.TestCase):
             # Too small for anything
             (-(2**63)-1, np.object),
             # Valid np.int64
-            (-(2**63)  , np.int64),
-            (-(2**63)+1, np.int64),
-            ((2**63)-1 , np.int64),
+            (-(2**63)  , np.int_),
+            (-(2**63)+1, np.int_),
+            ((2**63)-1 , np.int_),
             # Too big for np.int64, still valid np.uint64s
-            ((2**63)   , np.uint64),
-            ((2**63)+1 , np.uint64),
-            ((2**64)-1 , np.uint64),
+            ((2**63)   , np.uint),
+            ((2**63)+1 , np.uint),
+            ((2**64)-1 , np.uint),
             # Too big for anything
             ((2**64)   , np.object),
             ((2**64)+1 , np.object),
